LayoutTests/http/tests/iframe-monitor/shared-worker-expected.txt

CONSOLE MESSAGE: setup() finished: true
CONSOLE MESSAGE: Launched shared worker
CONSOLE MESSAGE: iframe is ready
CONSOLE MESSAGE: Launched shared worker
CONSOLE MESSAGE: iframe is ready
CONSOLE MESSAGE: Get message from parent window
CONSOLE MESSAGE: Send fetch request to worker
CONSOLE MESSAGE: Frame was unloaded because its network usage exceeded the limit.
CONSOLE MESSAGE: Frame was unloaded because its network usage exceeded the limit.
Test iframes using same shared worker are unloaded.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelector('iframe[name=frame1]').srcdoc is not ""
PASS document.querySelector('iframe[name=frame2]').srcdoc is not ""
PASS successfullyParsed is true

TEST COMPLETE
